dex/networks: validate secret size in contract
ExtractSwapDetails now verifies that the secret size is as expected.
It also ensures the secret size is encoded with an OP_DATA_1.

MakeContract now uses SecretKeySize instead of SecretHashSize, but they
were the same value previously.
Note that the caller of MakeContract should ensure their secret hash was
derived from a secret of length SecretKeySize.

TODO: SecretKeySize should be a common variable in a "dex" package, or
a per-asset setting, part of the dex.Asset struct, communicated between
client and server.