rpcserver: Add myorders route
While my orders were accessible through the exchanges endpoint, the data
was raw and unorganized. This adds some pre-computed, useful information
such as age of the order and amount settled. It also allows for
filtering by exchange and/or market.

* rpcserver: Remove orders from exchanges endpoint.

Because orders can now be accessed through the myorders endpoint, they
can be removed from exchanges.