client/{asset,core}: make order funding thread-safe
The client/asset order funding functions now keep a consistent view of
locked coins in both wallet and the funding coins map, preventing issues
with concurrent calls to e.g. FundOrder, FundingCoins, etc.

The btc backend now ensures funding coins are also tagged as "spendable"
in the listunspent result, ensuring the wallet has the keys to spend.

client/core: tweak logs and error messages

* client/asset/{btc,dcr}: fix locking issues

btc: resolve deadlock in FundOrder, no longer lock mtx in split()

dcr:
- Fix Swap unguarded lockFundingCoins call, also keep consistent view
- make unexported returnCoins that does not lock, for Swap.
- split() was not locking fundingMtx  before calling lockFundingCoins
- split() now keeps a consistent funding coin view from tx to locking
  funds and split funds

Note that dcr.split() is thread-safe, while btc.split() is not. This is
to accommodate the different usage patterns in FundOrder.
The approach in btc is slightly more efficient because the map and
wallet lock are only updated once regardless of split or not.
While in dcr the initial funding updates the map and locks in the
wallet, and does it again if split() is used.

* btc: enable wallet-fundingCoin map consistency check