server/{auth,market}: any request error is a miss
Count all errors from RequestWithTimeout in collectPreimages as a miss,
no longer trying to forgive the user based on unknown errors that
might indicate server connectivity trouble.
This simplifies the function, and avoids misuse of the piChan.

This also uses a new dex.ErrorKind in auth, ErrUserNotConnected,
to communicate to the caller of RequestWithTimeout where the error
occurred. This just helps logging for now.