client/asset/btc: sendrawtransaction syntax change
From node version 0.19, the sendrawtransaction RPC changed the
allowHighFees boolean argument to a maxFeeRate numeric argument.
btcd's rpclient detects which to use depending on the version obtained
from getnetworkinfo, however this detection is broken for nodes that
report a subversion that is not of the form "/Satoshi:0.20.1/". For
example, rpcclient fails to recognize "/LitecoinCore:0.20.1/" as
requiring the boolean argument.
To fix this, implement sendrawtransaction with rawrequest.
The (*ExchangeWallet).SendRawTransaction method will use the correct
argument type according to the interger network version (e.g. 200100)
detected on ExchangeWallet construction.