server/market: msgBook.{insert,update} for {book,update_remaining}
Make book new order updates use insert, and update remaining use update.
insert checks for preexistence of the entry and logs an error if it is
was already there. update checks and logs if it was not already there.
Both will assign to the map regardless, but these could be panics.

Fix routers_test.go send message synchronization.