server/market: fix unbook/update_remaining market-router sequence
* server/market: msgBook.{insert,update} for {book,update_remaining}

Make book new order updates use insert, and update remaining use update.
insert checks for preexistence of the entry and logs an error if it is
was already there. update checks and logs if it was not already there.
Both will assign to the map regardless, but these could be panics.

Fix routers_test.go send message synchronization.

* server/market: update_remaining BEFORE unbook

The update_remaining notes to the BookRouter must be sent before any
unbook notes for the same order, otherwise the BookRouter will remove
and then re-add the order when trying to update remaining.

Also change a couple type switches with one case to type assertions.