server/auth: connect resp excludes revoked match requests
The connect response includes a list of active matches, and it triggers
sending pending requests and messages. This excludes match-related
requests from being sent if the match in question is not included in the
list of active matches. Only pending revoke_match requests are allowed
in this case, purely as a courtesy to the client. Similarly, the connect
timeout for revoke_match requests is nown 4 hours.

Test RequestWhenConnected/SendWhenConnected.