client/core: auto-revoke matches when conditions warrant
revoke_match messages may be missed in unusual cases (such as poor connectivity
for extended periods). If the client app is not restarted, match negotiations
will continue for such matches (when connectivity is restored, for e.g.) and
this could cause a client to take their next action such as sending a swap even
though the match has been revoked.

Also, missing a revoke_match may cause other issues such as funding and change
coins remaining locked until client restart.

This adds a trackedTrade.maybeRevoke method to auto-detect matches that remain
in a status longer than expected, and with the expectation/assumption that such
matches would be revoked by the server, the client marks them as revoked.