client/core: safe swap and comms tweaks
client/core: prevent swap bcast with stale dex conn


Swap contracts are no longer permitted to be broadcast if the DEX
connection is known to be down. This can happen when a computer or
client process comes out of suspend and immediately tries to take their
action, particularly the taker trying to bcast their swap. The net
connection's read timeout must still put the link into a failed state
quickly for this to work during a resume, but this connectivity check is
also valuable with normal connectivity trouble, not just machine/process
suspend.

Triggering of ticks with the tick timer is also modified to skip tick
execution if a long delay since the previous timer fire is detected
(computer resume) to allow time for DEX re-connection. This together
with the swap check described above prevents needlessly broadcasting a
swap contract for a match that is revoked unbeknownst to the client. The
swap check is still necessary however since asset tip changes also
trigger a tick.

The (*wsConn).read loop now detects and logs timeouts immediately
before all other error checks to hasten flagging the connection as down
and to log the timeout as the cause in plain language.