core: fix some order book feed bugs
* fix book feed synchronization
* recognize epoch orders in initial order book
* don't add cancel orders to order book table

Moves responsibility for sending initial order book notification from 
webserver to core. Just populate the new order book feed with the 
initial FreshBookAction update as soon as it's created, under lock. This 
solves some bugs where the book is processed client side before 
book_order messages that came first, leading to double entries. This 
also precludes the fancy handling of the webserver.marketResponse that 
we were doing. That type is eliminated now, and the MarketPage handles 
the BookUpdate directly.

Include the epoch index in the OrderBook's epochQueue orders. The fixes 
a MarketPage bug where epoch orders were being added to the book, but 
were not recognized as epoch orders