suspend/resume
This wraps up the market suspend/resume functionality.

In summary, this is communicated in the following places:

- 'config' response, which includes market status for each market:

type MarketStatus struct {
	StartEpoch uint64 `json:"startepoch"`
	FinalEpoch uint64 `json:"finalepoch,omitempty"`
	Persist    *bool  `json:"persistbook,omitempty"` // nil and omitted when finalepoch is omitted
}

- The 'suspension' and 'resumption' routes sent by the server *at the time  
the status change happens*.  These are sent to 'orderbook' subscribers, and 
are sequenced with other book updates when appropriate.

- The 'suspension' and 'resumption' routes sent by the server *when the 
operator initially schedules it* via the administrative API.  These are 
_broadcasted_ notifications ahead of the actual event.

type TradeSuspension struct {
	MarketID    string `json:"marketid"`
	Seq         uint64 `json:"seq,omitempty"`         // only set at suspend time and if Persist==false
	SuspendTime uint64 `json:"suspendtime,omitempty"` // only set in advance of suspend
	FinalEpoch  uint64 `json:"finalepoch"`
	Persist     bool   `json:"persistbook"`
}

type TradeResumption struct {
	MarketID   string `json:"marketid"`
	ResumeTime uint64 `json:"resumetime,omitempty"` // only set in advance of resume
	StartEpoch uint64 `json:"startepoch"`
}

When the messages are sent at time of scheduling (in advance of the actual 
event), scheduled time is communicated for display to the user and recorded in 
the stored market config.

When the time roles around, the market uses the 'orderbook' subscription to 
mark the event at the right time and in sequence with other book note.  The 
'suspension' message received at this time is handled by the client by storing 
the FinalEpoch and Perist flags the in the relevant markets config thus 
marking it suspended (possibly redundant unless they missed the initial 
scheduling note), and if the `Persist` flag requires it, purging the order book. 
 The 'resumption' message received at time of resume updates MarketStatus 
with StartEpoch and clears FinalEpoch and Persist.

The 'config' request is no longer placed by the trade resumption handler since 
we have not implemented host market config changes with the server.  Plus it 
seems likely that the best way to handle this is to include the full 
msgjson.Market config for the resuming market *in* the trade resumption 
message at the time of resume.

Having these in the 'orderbook' subscription message sequence ensures the 
market is marked running/stopped at the proper time.  This replaces the timers.