Given the recently reduced importance of the redeem transactions being mined promptly, this introduces a new client exchange wallet option to set the confirmation target for the redeem transaction.
The contract transaction may already be created with segwit inputs to
lower fees (and accounting for this is corrected in #729), but the
redeem transactions will need p2wsh contract prevouts as their inputs to
get the segwit discounts as well. However, quickly mining the redeem
txns is not nearly important as with the swap contract txns. Thus, the
user should have the ability to use lower fee rates for the redeem txns.

The default for BTC is 2 blocks; default for DCR is 1 block.

This also sneaks in a test data race fix for server/asset/dcr, and
tweaks the swap tests as there were needlessly low timeouts in places.