chromium-based app
Look for a chromium browser, and if found, run it in app
mode pointing to the dexc webserver. Also create a system tray
menu entry so the user can reopen the window after closing it.

Create a tiny server to claim a port and prevent multiple windowed
instances of dexc running at the same time. Try to hit the server
port on startup and bail on success. The server will signal when a
new connection is received. The signal can be used to "restore focus"
to the app window. Afaict, actually bringing the window to the front
is not possible since we are starting chromium with exec, but if the
user attempts to start a second instance of dexc, close the old window
and open a new one to simulate focus. I acknowledge that this may
not always be the best UX.

If the client has active orders when the user tries to shutdown,
open a simple warning/confirmation dialog using
github.com/gen2brain/dlgs.