server/{db,swap},client/core: no maker redemption ack, no redeem confs
server/db

- Remove aSigAckOfBRedeem column from matches table.

- Storing the participant (taker) redeem coin sets active = FALSE. 
Previously this was set when both redemption acks were received.

Now any match in the table with status MatchComplete will also be 
inactive. Previously matches could be (and often were) MatchComplete but 
also active. This was not great.

server/swap

- processBlock no longer checks the redeem txns for their confirmations. 
We were checking against swapConf, but confirmations on the redeem txns 
never were important, only that both parties could author and bcast 
their redeem txns (the secret was shared and both parties used it to 
spend the respective contract outputs).

- processBlock just checks swap txns against swapConf, and it now 
unlocks the order funding coins for the order when the swap txn that 
spent them reaches swapConf (previously it waited to do this when the 
entire match was completed). This reflects the "event-based" and 
"block-based" inaction check distinction.

- processRedeem is now the spot where a match is deleted, when the 
taker's redeem is validated (when MatchComplete and active=FALSE are 
set). Note that this means there will no longer be matches in Swapper 
memory with MatchComplete status.

- processRedeem now does the cancellation rate accounting that was 
previously done in processAck since it is the 'redeem' validation that 
marks the end of the match for the user, not the redemption ack.

- processRedeem no longer sends a 'redemption' request to the maker.

- processAck no longer processes 'redemption' acks from the maker.

client/core

- Maker no longer requires 'redemption' request following taker's redeem 
to move to MatchComplete.

- Maker is only MakerRedeemed if their 'redeem' request following the 
actual broadcast fails, otherwise they jump straight to MatchComplete. 
This is switched by auth.RedeemSig.

- Fix some calls references to MatchProof.IsRevoked in logging 
statements not being called as a function.