client/core: fix resendPendingRequest sending wrong redeem coin
The coins sent with the retried 'redeem' requests were empty. This fixes
the finalizeRedeemAction call to use the redeem coin.

This also removes a misplaced log in resolveServerMissedTakerRedeem.
It is not unexpected for the maker to be at MatchComplete with the
server at MakerRedeemed if the taker hasn't done their 'redeem'.

In trade_simnet_test.go, set Core's logger in core.Config.  Note that
these simnet tests are presently failing due to recent changes on master.