assets: implement segwit btc swap contracts
This implements segwit as an option for the wallet backends and
server's asset backends. Bitcoin is configured to use segwit by
defualt. Litecoin likely will too, and I've tested it, but I'm
leaving it in non-segwit just to demonstrate the option. It's
important to offer the option, because even though some bitcoin
forks won't have segwit, our clone backend might still be able to
accommodate.

* fix dust check and test dust edges with both dcr and btc