market: verify unspent status of unfilled orders
When a user submits a new order, all of that user's unfilled book orders 
have their funding coins checked. If any of them are spent, those orders 
are unbooked. The newly submitted order is processed normally after 
that. The idea behind checking on order placement is to prevent a user 
from moving the same bunch of coins around into new utxos to fund a 
large number of bogus (unfunded) orders. Although a user could submit 
multiple orders with valid funding coins, then proceed to spend those 
coins, this user could affect similar disruption by simple not executing 
swaps on match, so continually checking funding coins for online users 
is not particularly useful

This requires spending txns to be mined, not just in mempool, before the 
check detects the funding coins as spent. This is the same with the 
check in NewMarket

The check uses gettxout (for btc) on each coin ID, but a better check 
could involve pulling the blocks since the previous check and checking 
the block contents against a list of coins of interest. DCR could use 
cfilters. More complex, asset specific.

Perhaps the best approach would be to register coin IDs with the 
backends so the backends watch and notify, but that's even more complex.