server/{auth,db,market,swap}: user order quantity limits
This implements a concept for limiting user order quantities.

- Users are limited in the amount they can have in active swaps 
("settling"). The limit is computed based on the user's match history.

- Only orders that are considered "likely taker" are subject to this 
limit. Likely taker includes: market order, limit with TiF immediate, 
limit with TiF standing and a rate that is within the spread.

- Separate to the taker order limit, there is an constant limit for 
booked order quantity in lots, to which all submitted standing limit 
orders are subject to. To be conservative, an order may be considered 
"likely taker" AND be subjected to this book limit at the same time.

- Match outcome history now includes value of the match (swapped, stuck 
long, stuck short, and spoofed). The match outcome history depth is the 
same as the scoring system's depth (same data structure presently).

- Users are granted an init/base limit that is defined in number of 
lots. The market converts this to an amount based on the current lot 
size.

- The quantity of new orders are validated according to the user's 
current active amount (booked + swapping in the market) and their active 
amount limit (determined by historical match outcomes for the market and 
the base limit).

- Limits, active amounts, and match history are per-market values.

- The user only finds out if they hit the limit when submitting the 
order. This may be inconvenient when tx split is enabled as it creates 
an on-chain transaction. Server could provide the user's current limit 
in the connect response, and the client could track their limit based on 
the formula.