add market data API endpoints
Exposes a rate-limited HTTP API. Data endpoints are accessible via HTTP
or WebSockets. The result for the WebSocket request is identical to the
REST API response.

Three type of data are available.

/spots is the spot price and booked volume of all markets.

/candles is candlestick data, available in bin sizes of 24h, 1h, and 15m.
An example URL is /candles/dcr/btc/15m.

/orderbook is already a WebSocket route, but is now also accessible by HTTP.
An example URL is /orderbook/dcr/btc

/config is another WebSocket route that is also now available over HTTP too.

The data API implements a data cache, but does not cache pre-encoded
responses for /candles or /orderbook (yet).