client/core: handle cancel order for trades with reconciled status
If a trade and a cancel order targeting that order were in epoch status, 
and order status reconciliation happened, the cancel order would stay in 
epoch status forever.

This updates (*dexConnection).reconcileTrades so that when a trade order 
gets its status updated (EDIT: to a status other than canceled), any 
associated cancel order is unlinked and put into revoked status 
(retired) to it is not loaded forever on startup.

This also updates (c *Core).dbTrackers to retire cancel orders loaded 
from the DB that have no associated active trade. This is a "should not 
happen" case, but which I hit when testing unrelated server changes.

This also moves the final log statement in dexc's main.go to before the 
log rotator closes.