core,ui: add confirmation for currently waiting swap
Add confirmation count to match cards on the order details page, for 
matches in MakerSwapCast or TakerSwapCast. It's a bigger job than it 
sounds like, because 1) we weren't tracking confirmation count on our 
own swaps yet, and 2) I didn't think we should add the count unless it 
updated live, so the count, match status, and various coin ID links will 
update live now.