testing: add a load tester
The loadbot package provides "programs" for load testing the simnet harness.
Non-ideal network conditions can be simulated with the use of a proxy server.
The bot will quit as soon as an error is encountered while trying to place an order.

There are three primary programs. You run one program at a time.

./loadbot -p pingpong
./loadbot -p sidestacker
./loadbot -p compound

Debug logging can be enabled with the -debug flag. Trace logging can be enabled with the -trace flag.

See the loadbot README for network condition simulation options.