client/multi: add max order estimation
Get maximum order estimations for a given market and side based on the
wallet balance and, in the case of buy orders, the rate in the rate input field.
The data is sourced directly from the wallet, because the limits change
based on available funding coins.

The data is shown in the UI as a small message above the rate field. When
you click on the label, the quantity fields are pre-populated with the max order.