multi: add market data API endpoints
Exposes a rate-limited HTTP API. Data endpoints are accessible via HTTP 
or WebSockets. The result for the WebSocket request is identical to the 
REST API response.

/spots is the spot price and booked volume of all markets.

/candles is candlestick data, available in bin sizes of 24h, 1h, 15m, 
and per-epoch sticks. An example URL is /candles/dcr/btc/15m.

/orderbook is already a WebSocket route, but is now also accessible by 
HTTP. An example URL is /orderbook/dcr/btc

/config is another WebSocket route that is also now available over HTTP 
too.

The data API implements a data cache, but does not cache pre-encoded 
responses for /candles or /orderbook (yet).

**server/db**

Market history is stored as a row per epoch. I've created a new table 
for this, though we could also combine with the existing epoch table 
that holds match proof info if preferred. 

I've made no attempt to integrate pre-existing data. Data will be 
collected from the time of upgrade only. The upgrade is silent, since we 
don't have upgrade infrastructure in place yet for the server. 

**server/comms**

The config route is now registered as an http handler, but also called 
from the websocket message handler.

HTTP requests are rate-limited using `"golang.org/x/time/rate"`. There 
is both a global rate limiter and a per-IP rate limiter. The rate limits 
apply to un-authenticated routes only. The /config request is has an 
exception as a "critical' route.

* enable/disable via admin api. config route not metered on ws

* use better http json response encoding pattern