client/core: avoid potential wallet reconfigure deadlock
This prevents a deadlock on wallet reconfigure that would happen if the 
client asset backend's tip change handler (a Core method) attempted to 
acquire the Core.walletMtx because it was already locked in 
ReconfigureWallet. The deadlock was on oldWallet.Disconnect inside 
ReconfigureWallet because it would be waiting for the block monitor 
goroutine to return that was likely to itself be waiting on Core's 
walletMtx (e.g. monitorBlocks>checkForNewBlocks>tipChange).