order fee estimation
An order fee estimate consists of 3 values for the swap and 2 for
the redemption. The swap has a max possible fees (MaxFeeRate),
and both high and low estimates corresponding to the best and worst
case settlement sequences, but at the prevailing network fee rate.

Similarly, the redemption estimate has high and low estimates
evaluated for the worst and best settlemetn sequences (1 tx per lot
vs 1 tx for all).