testing: Add Ethereum simnet harness
This adds a basic Ethereum private network testing harness.

This network relies on proof of authority. More about this can be found
at https://eips.ethereum.org/EIPS/eip-225

It creates two nodes, either of which can add new blocks to the network.
Each node starts with funds that can be sent to addresses that we
later create in tests.

The harness does not start an rpc client. I think it is best that we use ipc
(inter-process communication) as this seems to be the most secure.
Other options are an http or ws server, but neither is permissioned.
They can be restricted to the local network, but at that point it doesn't
seem much different than using ipc.