server/market: suspend with purge needs to clear BookRouter
Although market suspend with persist=false was purging the book both in
Market's in-memory book and in the DB, the outgoing BookRouter's msgBook
was not resetting its orders map.

This change reinitializes the msgBook.orders map on purge.

This also moves a bit of repetative code out of the main runBook loop.