multi: send raw tx data in audit and match_status requests
This implements tx data in the audit and match_status requests.

To accommodate SPV clients, who don't have a mempool and so are in the dark
until a contract's transaction is mined, we are going to provide the serialized swap
transactions to the clients. The server sends the data in the msgjson.Audit payload.
Server also sends the info in the msgjson.MatchStatusResult, but only for
status/side combos for which it is relevant.

The wallet does not make use of the info yet, but it does receive the data though
the modified (Wallet).AuditContract method. There are some other changes that
will be needed in the way Core handles the audit too, but I think it'll be easier to
handle that during work on the wallet.