client/comms: wsConn.Connect should trigger keepAlive on error
A DEX connection attempt must start the reconnect loop if it
fails to connect initially.

Allow messages on reconnect by starting c.listen(dc) when
the wsConn is active, even waiting to (re)connect.

Redo the handleConnectEvent to it triggers on initial connect too.

handleReconnect must work for locked or read-only conns, so make
the authDEX call conditional.