client/{asset,core}: do not disconnect wallets in setWalletPassword
When an asset.Wallet is disconnected, it may not be possible to
reconnect it.  Document this with the asset.Wallet interface.

Always leave wallets connected in setWalletPassword. Callers like
SetWalletPassword have the xcWallet in the wallets map, so it should
stay live and usable.

In ReconfigureWallet, call connectWallet before setWalletPassword
anyway because the connection approach used by setWalletPassword
was too involved for the fresh xcWallet created by ReconfigureWallet
that is not yet in Core's wallets map.