multi: no Seq in msgjson.EpochReportNote
Since there's no book or order updates with epoch_report, it should not
increment Seq. Having seq incremented in a message route that 0.1.x
client's don't understand also breaks compatibility with 0.2 servers.

This also updates the docs for seq in the server/market.subscribers
struct to reflect the purpose of the Seq value as per the spec:

    The order book and all updates include a sequence ID, which
increments by +1 whenever the DEX accepts, removes, or modifies an
order. The client is responsible for tracking the sequence ID to ensure
all order updates are received. If an update appears to be missing, the
client should re-subscribe to the market to synchronize the order book
from scratch.