client/core: on shutdown, log active orders and unlock retired trade coins
Usually the ticker retires orders, which also returns any coins that may
still be locked in the wallets for such orders. This supplements the
shutdown of (*Core).Run with logging of active orders, and unlocks any
such coins for retired orders that the ticker has not yet processed.

Note that if any coins are still locked for inactive orders, that
indicates a bug elsewhere since we should have returned these coins when
the order reached an inactive status. This should help debugging.