client/core: marketTrades returns only active trades
The User and Exchanges methods of Core include orders slices, which have
thus far included all orders including thoes that are inactive and will
be retired on the next tick. This changes marketTrades to only return
active trades. Now Logout's handling of active orders is consistent with
the orders slice returned by User or Exchanges.