reimagine orderbook.OrderBook.orders map
The orders map is now just stores the rate and side for an order ID.
The side field specified which bookSide it would be in, while rate
facilitates efficiently locating the order in a bookSide by specifiying
which rate bin it belongs in.

Provide rateBin to both bookSide.Remove and UpdateRemaining to avoid
having to search for the bin. Remove already looked up the bin by rate,
but now Remove gets the OrderID and rate directly, rather than from
Order fields.  Passing rate to UpdateRemaining saves an outer loop,
and it mirrors how Remove accesses an order.

(*bookSide).UpdateRemaining now clones the order being modified so
that copies that were previously retrieved from Orders() or BestNOrder()
do not have their Quantity field modified unsafely.