client: fix db store of enc pass in ReconfigureWallet
When ReconfigureWallet is used to change the wallet's passphrase, there
are two DB updates, but the second one was using the original
EncryptedPW.  While two DB writes is inefficient, this still uses the
(*Core).setWalletPassword method as is since it also checks the password
and emits a notification.

The resolution is to simply update the EncryptedPW field of the
client/db.Wallet instance passed to db.UpdateWallet in ReconfigureWallet
after updating the wallet balance.

This also revises (*BoltDB).SetWalletPassword so that it does not bother
to decode the current wallet's db.Balance since only the wallet encoding
bucket (walletKey) is updated, not the balance bucket (balanceKey).
A deep copy of the new encrypted PW is also done out of caution.