server: ensure all fee rates are capped by config
Both BoookRouter and OrderRouter use FeeManager.LastRate to obtain
fee rates, and these need to have the configured MaxFeeRate applied.

The rate FeeFetcher.FeeRate is also now capped, but consumers scale this
and ally the limit after scaling anyway. (*Market).processReadyEpoch
already applied the MaxFeeRate after scale, but  (*Server).apiAsset was
neglecting this cap and it is now added.

This applies to the rates returned by the 'orderbook' and 'fee_rate'
respones from the Book Router.

This also applies to the zero-conf funding coin check performed by the
OrderRouter in checkZeroConf. Previously zero-conf coins were
held to a different and potentially higher fee rate than swaps that
could exceed MaxFeeRate.