multi: asset versioning
This adds asset version to the client and server. The config response includes the
asset version for each asset. The client checks the asset version and blocks orders
if it does not support the server's asset version.

When a new asset version is added in the future, the client should support both
the new and previous version unless there are security issues with the old one.
See "MasterWallet" discussion.